+++ module-application/src/main/java/com/zoopi/exception/BusinessException.java
@@ -3,8 +3,8 @@ package com.zoopi.exception;
 import java.util.ArrayList;
 import java.util.List;
 
-import com.zoopi.model.ErrorCode;
-import com.zoopi.model.ErrorResponse;
+import com.zoopi.ErrorCode;
+import com.zoopi.ErrorResponse;
 
 import lombok.Getter;
