+++ module-application/src/main/java/com/zoopi/exception/EntityNotFoundException.java
@@ -1,6 +1,6 @@
 package com.zoopi.exception;
 
-import com.zoopi.model.ErrorCode;
+import com.zoopi.ErrorCode;
 
 public class EntityNotFoundException extends BusinessException {
