+++ module-application/src/main/java/com/zoopi/security/exception/CustomAuthenticationException.java
@@ -5,8 +5,8 @@ import java.util.List;
 
 import org.springframework.security.core.AuthenticationException;
 
-import com.zoopi.model.ErrorCode;
-import com.zoopi.model.ErrorResponse;
+import com.zoopi.ErrorCode;
+import com.zoopi.ErrorResponse;
 
 import lombok.Getter;
