+++ module-application/src/main/java/com/zoopi/security/jwt/exception/JwtAuthenticationException.java
@@ -3,8 +3,8 @@ package com.zoopi.security.jwt.exception;
 import java.util.List;
 
 import com.zoopi.security.exception.CustomAuthenticationException;
-import com.zoopi.model.ErrorCode;
-import com.zoopi.model.ErrorResponse;
+import com.zoopi.ErrorCode;
+import com.zoopi.ErrorResponse;
 
 import lombok.Getter;
