--- module-application/src/main/java/com/zoopi/service/certification/BloodDonationCertService.java
@@ -1,14 +1,0 @@
-package com.zoopi.service.certification;
-
-import org.springframework.stereotype.Service;
-
-import lombok.RequiredArgsConstructor;
-
-import com.zoopi.domain.certification.repository.BloodDonationCertRepository;
-
-@Service
-@RequiredArgsConstructor
-public class BloodDonationCertService {
-
-	private final BloodDonationCertRepository certRepository;
-}
