+++ module-model/src/main/java/com/zoopi/client/model/member/JwtDto.java
@@ -1,0 +1,13 @@
+package com.zoopi.client.model.member;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@Getter
+@AllArgsConstructor
+public class JwtDto {
+
+	private String accessToken;
+	private String refreshToken;
+
+}
