+++ module-model/src/main/java/com/zoopi/client/model/member/SigninResponse.java
@@ -1,0 +1,19 @@
+package com.zoopi.client.model.member;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@Getter
+@AllArgsConstructor
+public class SigninResponse {
+
+	private JwtDto jwt;
+	private SigninResult result;
+
+	public enum SigninResult {
+
+		SUCCESS, NONEXISTENT_USERNAME, MISMATCHED_PASSWORD
+
+	}
+
+}
