--- module-model/src/main/java/com/zoopi/model/client/member/JwtDto.java
@@ -1,13 +1,0 @@
-package com.zoopi.model.client.member;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@Getter
-@AllArgsConstructor
-public class JwtDto {
-
-	private String accessToken;
-	private String refreshToken;
-
-}
